+++ src/main/java/com/rfchina/wallet/server/api/impl/SeniorPayApiImpl.java
@@ -110,7 +110,7 @@ public class SeniorPayApiImpl implements SeniorPayApi {
 		@ParamValid(nullable = false) Long walletId,
 		@ParamValid(nullable = false) Long cardId,
 		@ParamValid(nullable = false) Long amount,
-		@ParamValid(nullable = false) Byte validateType,
+		@ParamValid(nullable = false) Integer validateType,
 		@ParamValid(nullable = false) String jumpUrl,
 		@ParamValid(nullable = false) String customerIp) {
 
@@ -122,7 +122,7 @@ public class SeniorPayApiImpl implements SeniorPayApi {
 		}
 
 		WithdrawResp withdraw = seniorPayService
-			.withdraw(walletId, walletCard, amount, validateType, jumpUrl, customerIp);
+			.withdraw(walletId, walletCard, amount, validateType.byteValue(), jumpUrl, customerIp);
 
 		UnifiedConfirmVo confirmVo = BeanUtil.newInstance(withdraw, UnifiedConfirmVo.class);
 		confirmVo.setOrderId(withdraw.getId());
