+++ src/main/java/com/rfchina/wallet/server/web/SeniorPayController.java
@@ -83,7 +83,7 @@ public class SeniorPayController {
 		@ApiParam(value = "钱包id", required = true) @RequestParam("wallet_id") Long walletId,
 		@ApiParam(value = "银行卡id", required = true) @RequestParam("card_id") Long cardId,
 		@ApiParam(value = "金额", required = true) @RequestParam("amount") Long amount,
-		@ApiParam(value = "交易验证方式 0：无验证 1：短信 2：密码") @RequestParam("validate_type") Byte validateType,
+		@ApiParam(value = "交易验证方式 0：无验证 1：短信 2：密码") @RequestParam("validate_type") Integer validateType,
 		@ApiParam(value = "跳转地址", required = true) @RequestParam(value = "jump_url") String jumpUrl,
 		@ApiParam(value = "客户Ip", required = true) @RequestParam(value = "customer_ip") String customerIp
 	) {
